Crate node_sys

Raw bindings to the Node.js API for projects using wasm-bindgen

Modules

assert
The assert module provides a set of assertion functions for verifying invariants.

async_hooks
The async_hooks module provides an API to register callbacks tracking the lifetime of asynchronous resources created inside a Node.js application.

child_process
The child_process module provides the ability to spawn child processes in a manner that is similar, but not identical, to popen(3).

cluster
The cluster module allows easy creation of child processes that all share server ports.

console
The console module provides a simple debugging console that is similar to the JavaScript console mechanism provided by web browsers.

crypto
The crypto module provides cryptographic functionality that includes a set of wrappers for OpenSSL's hash, HMAC, cipher, decipher, sign, and verify functions.

dgram
The dgram module provides an implementation of UDP Datagram sockets.

dns
The dns module contains functions for performing name resolution either through the OS or by connecting to a name server.

events
The events module provides functionality for working with EventEmitter.

fs
The fs module provides an API for interacting with the file system in a manner closely modeled around standard POSIX functions.

http
The http module provides an implementation of the HTTP protocol for clients and servers.

http2
The http/2 module provides an implementation of the HTTP/2 protocol for clients and servers.

https
The https module provides an implementation of HTTP over TLS/SSL for clients and servers

inspector
The inspector module provides an API for interacting with the V8 inspector.

net
The net module provides an asynchronous network API for creating stream-based TCP or IPC servers (net.createServer()) and clients (net.createConnection()).

os
The os module provides operating system-related utility methods and properties.

path
The path module provides utilities for working with file and directory paths.

perf_hooks
The perf_hook module provides an implementation of the W3C Performance Timeline specification.

querystring
The querystring module provides utilities for parsing and formatting URL query strings.

readline
The readline module provides an interface for reading data from a Readable stream (such as process.stdin) one line at a time.

repl
The repl module provides a Read-Eval-Print-Loop (REPL) implementation that is available both as a standalone program or includible in other applications.

stream
The stream module provides an API for implementing the stream interface.

string_decoder
The string_decoder module provides an API for decoding Buffer objects into strings in a manner that preserves encoded multi-byte UTF-8 and UTF-16 characters.

timers
The timer module exposes a global API for scheduling functions to be called at some future period of time.

tls
The tls module provides an implementation of the Transport Layer Security (TLS) and Secure Socket Layer (SSL) protocols that is built on top of OpenSSL.

trace_events
The trace_event module provides functionality to centralize tracing information generated by V8, Node.js core, and userspace code.

tty
The tty module provides the tty.ReadStream and tty.WriteStream classes.

url
The url module provides utilities for URL resolution and parsing.

util
The util module is primarily designed to support the needs of Node.js' own internal APIs.

v8
The v8 module exposes APIs that are specific to the version of V8 built into the Node.js binary.

vm
The vm module provides APIs for compiling and running code within V8 Virtual Machine contexts.

worker_threads
The worker_threads module enables the use of threads that execute JavaScript in parallel.

zlib
The zlib module provides compression functionality implemented using Gzip and Deflate/Inflate, as well as Brotli.

Structs

BeforeExitListener

Buffer

BufferModule

CallSite

ConsoleConstructor

ConsoleConstructorOptions

CpuUsage

DisconnectListener

Domain

ErrnoException

EventEmitter
An object that emits events which trigger functions registered as listeners.

ExitListener

Global

HrTime

InspectOptions

MemoryUsage

MessageListener

Module

MultipleResolveListener

MultipleResolveType

NewListenerListener

NodeModule

NodeRequireFunction

Platform

Process

ProcessEnv

ProcessFeatures

ProcessRelease

ProcessReport

ProcessSendOptions

ProcessVersions

ReadWriteStream

ReadableStream

RejectionHandledListener

RemoveListenerListener

RequireResolve

Signals

SignalsListener

Socket

Timeout

Timer

TypedArray

UncaughtExceptionListener

UnhandledRejectionListener

WarningListener

WriteableStream

Statics

buffer

global

process

require

Functions

transcode

Type Definitions

TranscodeEncoding